And the new memory reader is out the window.

+++ d3x/Structures/AttributeGroupsContainer.cs
@@ -13,7 +13,7 @@ namespace D3x.Structures
         public Wtf b; // 0x01C 
         public Wtf c; // 0x038 
         public Wtf d; // 0x054 
-        public IntPtr FastAttribGroups;
+        public UInt32 FastAttribGroups;
         //public Container<CAttribGroup>* FastAttribGroups; // 0x070 
         [MarshalAs(UnmanagedType.ByValArray, SizeConst = 32)]
         public byte[] unknown_74; // 0x074
